backend/module_1_techwatch/requirements.txt
feedparser
requests
beautifulsoup4
trafilatura
psycopg[binary]
pyyaml
qdrant-client>=1.17.0
jinja2
sentence-transformers
google-generativeai
schedule
ddgs
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0
fastapi
uvicorn
python-multipart